# contracts/http/openapi.yaml
...
openapi: 3.1.0
info:
  title: Research Radar — HTTP wire contract
  version: 0.3.0
  summary: 'Hợp đồng HTTP cho toàn bộ operation `transport: http` của contracts/ports.yaml.'
  description: |-
    Đây là hợp đồng dây (wire) của Research Radar, sinh từ `contracts/ports.yaml` và giữ nguyên `operation_id`
    làm `operationId`. Mỗi operation `transport: http` trong ports.yaml có ĐÚNG MỘT cặp path+method ở đây.

    Những gì file này KHÔNG định nghĩa, và nơi chúng thuộc về:
      - Ngữ nghĩa transition trạng thái  → contracts/state/*.yaml (PC03)
      - Danh mục mã lỗi và envelope      → contracts/errors.yaml (PC03); `ErrorEnvelope` dưới đây là bản sao ĐÚNG
                                           hình dạng của `error_envelope` trong file đó
      - Ngân sách retry/timeout/lease    → contracts/retry-policy.yaml (PC03)
      - Ranh giới module và cạnh bị cấm  → contracts/modules.yaml (PC01)
      - Nội dung/coverage báo cáo        → contracts/reporting/* (PC04)
      - Task và schema AI                → contracts/ai/* (PC06)
      - Ngữ nghĩa lệnh Telegram          → contracts/telegram/commands.yaml (PC07)
      - Vòng đời secret, CSRF, xoay token→ contracts/ops/secrets.md (PC08)

    Wire rules bắt buộc (SRC-PLAN §5.1), áp cho mọi operation:
      - ID truyền dưới dạng string; timestamp UTC RFC 3339 với độ chính xác mili giây (AMD-B08).
      - Mutation từ worker mang `request_id`, `schema_version`, `idempotency_key`, `payload_hash`; mutation đang
...

...
  x-contract:
...
    - REQ-D04
...
servers:
...
security:
- ownerSessionCookie: []
  ownerCsrfToken: []
components:
  securitySchemes:
    ownerSessionCookie:
      type: apiKey
      in: cookie
      name: rr_session
      description: 'Cookie phiên HttpOnly + Secure + SameSite=Lax. PROVISIONAL, phối hợp với PC08 (xem info.x-contract.security_scheme_decision).
        Với MUTATION, scheme này xuất hiện CÙNG `ownerCsrfToken` trong MỘT security requirement — tức quan hệ AND, cookie
        một mình KHÔNG đủ. Với read-only, cookie đứng một mình: CSRF bảo vệ thao tác đổi trạng thái, không phải phép đọc.'
    ownerCsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 'CSRF double-submit: giá trị phải khớp cookie `rr_csrf` (không-HttpOnly). Bắt buộc cho MỌI mutation của
        owner session. Thiếu hoặc lệch ⇒ 403 `CSRF_REJECTED` (KHÔNG phải `FORBIDDEN_EDGE`: cạnh vẫn hợp lệ, chỉ lời gọi
        không chứng minh được nó do trang của ứng dụng phát ra — ruling CR-PC08-03/CR-PC08-04). Phiên KHÔNG bị hủy khi gặp
        mã này.'
    collectorToken:
      type: http
      scheme: bearer
      description: Token của collector, lưu trong file cấu hình QUYỀN HẠN CHẾ trên máy cá nhân, KHÔNG BAO GIỜ trong repo
        (SRC-SPEC §11.2). Chỉ mở được các operation `worker.*` và `ingest.*`; gọi `save.create` bằng token này ⇒ 401 (denied
        case NC-01).
    analysisWorkerToken:
      type: http
      scheme: bearer
      description: Token của analysis worker. Chỉ mở `analysis.*` và `secret.issue_task_credential`. KHÔNG mở đường tới
        tag, report hay Saved (denied case NC-02).
    telegramIngressSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Bí mật webhook do app đặt khi đăng ký webhook với Telegram. Sai hoặc thiếu ⇒ 401 và KHÔNG xử lý update.
        Biết chat ID KHÔNG phải là quyền (SRC-SPEC §11.3).
    backupOperatorToken:
      type: http
      scheme: bearer
      description: 'Danh tính operator backup. Tách khỏi phiên owner vì restore là thao tác có khóa side effect (AMD-B11).
        Sửa ở FIX1 (CR-PC08-02): mọi route `backup.*` nhận DUY NHẤT scheme này — trước đây phiên trình duyệt cũng qua được,
        nghĩa là một tab đang mở có thể kích hoạt restore.'
  parameters:
    SchemaVersionHeader:
      name: X-Schema-Version
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/SemVer'
      description: Phiên bản wire schema của request. KHÔNG dùng 'latest' (SRC-PLAN §5). Lệch major ⇒ 422 `VALIDATION_ERROR`
        với `details_safe.violation_kind = schema_version_unsupported`.
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/Ulid'
      description: ID của LẦN GỌI này. Một retry sinh `X-Request-Id` MỚI nhưng GIỮ NGUYÊN `Idempotency-Key` — đó là cách
        server phân biệt replay với một yêu cầu mới.
...
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      description: |-
        Envelope lỗi chung (SRC-PLAN §5.1). Hình dạng này là BẢN SAO của `error_envelope` trong contracts/errors.yaml;
        nếu hai bên lệch nhau thì contracts/errors.yaml THẮNG và file này phải sửa.

        `correlation_id` LUÔN có mặt, kể cả khi `storage.health != healthy`: nó sinh trong bộ nhớ tiến trình, không cần
        một hàng DB. Không mã lỗi nào đòi hỏi ghi được vào DB mới coi là đã báo (contracts/errors.yaml §EPR-01).

        Envelope KHÔNG BAO GIỜ chứa: transcript của model, prompt gốc, nội dung post nguyên văn, API key, bot token,
        cookie, header Authorization, chat ID, đường dẫn file tuyệt đối, stack trace, câu SQL, tên bảng/cột nội bộ.
      required:
      - code
      - scope
      - retry_class
      - message_safe
      - correlation_id
      properties:
        code:
          type: string
          enum:
          - X_CHALLENGE_REQUIRED
          - X_ACCESS_BLOCKED
          - WORKER_LEASE_EXPIRED
          - STALE_LEASE
          - INGEST_ACK_LOST
          - IDEMPOTENCY_CONFLICT
          - SOURCE_METADATA_UNAVAILABLE
          - IDENTITY_CONFLICT
          - AI_OUTPUT_INVALID
          - AI_PROVIDER_UNAVAILABLE
          - AI_ATTEMPT_UNCERTAIN
          - EMBEDDING_GENERATION_MISMATCH
          - TAG_VERSION_STALE
          - CONFLICT
          - TELEGRAM_SEND_UNCERTAIN
          - TELEGRAM_PERMANENT_FAILURE
          - UNAUTHORIZED_COMMAND
          - STORAGE_WRITE_FAILED
          - RESTORE_UNVERIFIED
          - SOURCE_LAYOUT_CHANGED
          - CSRF_REJECTED
          - UNAUTHORIZED
          - CAPABILITY_DENIED
          - FORBIDDEN_EDGE
          - VALIDATION_ERROR
          - NOT_FOUND
          - RATE_LIMITED
          - INTERNAL
          description: Phải là một mã đã khai trong contracts/errors.yaml. Không có mã ngoài danh mục.
        scope:
          type: string
          enum:
          - run
          - item
          - delivery
          - storage
          - request
          description: Bằng đúng `scope` của mã trong contracts/errors.yaml.
        retry_class:
          type: string
          enum:
          - none
          - retryable_with_budget
          - needs_user
          - operator_decision
          - unknown_outcome
          description: Bằng đúng `retry_class` của mã. `unknown_outcome` nghĩa là KHÔNG BIẾT side effect đã xảy ra hay chưa
            — client KHÔNG được coi là rollback.
        message_safe:
          type: string
          maxLength: 300
          description: Tiếng Việt, render từ `message_safe_template` của mã. KHÔNG chèn giá trị ngoài `details_safe_keys`.
        correlation_id:
          $ref: '#/components/schemas/Ulid'
        details_safe:
          type:
          - object
          - 'null'
          description: Chỉ chứa các khóa liệt kê ở `details_safe_keys` của mã trong contracts/errors.yaml. Khóa lạ ⇒ producer
            tự từ chối, không gửi ra.
        retry_after_ms:
          type:
          - integer
          - 'null'
          minimum: 0
          description: Chỉ có mặt khi `retry_class = retryable_with_budget` hoặc mã `RATE_LIMITED`. null nghĩa là KHÔNG
            có gợi ý, KHÔNG phải 'retry ngay'.
    Ulid:
      type: string
      pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
      description: ID truyền dưới dạng string (SRC-PLAN §5.1).
    SemVer:
      type: string
      pattern: ^[0-9]+\.[0-9]+\.[0-9]+$
...
    PageLimit:
      type: integer
      minimum: 1
      maximum: 100
      default: 25
      description: Phân trang bằng cursor; `limit` có trần cứng 100 để một response không phình.
...
paths:
...
  /v1/reports:
    get:
      operationId: report.list
      summary: Danh sách báo cáo theo kỳ (màn hình mặc định)
      description: 'Phân trang.


        Response mang `report_build_id` (UUIDv4, `UNIQUE(owner_id, report_build_id)` — khóa idempotency của `report.publish`,
        F-A1R2-01) và `abort_reason` (enum `empty_period | tag_version_stale | embedding_generation_mismatch | cas_conflict
        | builder_failure | cancelled`, NOT NULL khi và chỉ khi `status=''aborted''`). Hai cột do PC02 FIX3 khai trên ENT-report
        theo bảng field-level của ruling FIX3; hình dạng chi tiết ở `contracts/schemas/report.schema.json` (PC04). `abort_reason=''empty_period''`
        là một kỳ RỖNG HỢP LỆ, KHÔNG phải lỗi — UI không được gộp nó với `builder_failure` (I13).'
      tags:
      - reports
      security:
      - ownerSessionCookie: []
      x-auth-scope: owner_session
      x-owner-module: MOD-report-service
      x-caller-modules:
      - MOD-web-ui
      x-mutation: false
      x-scenario-refs:
      - SC08
      - SC15
      x-requirement-refs:
      - REQ-D45
      - REQ-P0-07
      x-decision-refs:
      - B04
      x-invariant-refs:
      - I06
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/PageLimit'
        description: Số report tối đa.
      - name: cursor
        in: query
        required: false
        schema:
          type: string
        description: Con trỏ trang.
      - $ref: '#/components/parameters/SchemaVersionHeader'
      - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: Ngày, coverage `[from, to)`, số mục, số hướng nổi, quality.
          content:
            application/json:
              schema:
                $ref: ../schemas/report.schema.json
          headers:
            X-Schema-Version:
              description: Phiên bản schema của response.
              schema:
                $ref: '#/components/schemas/SemVer'
        '401':
          description: 'Lỗi: `UNAUTHORIZED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 'Lỗi: `FORBIDDEN_EDGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: 'Lỗi: `INTERNAL`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/reports/{report_id}:
    get:
      operationId: report.get
      summary: Chi tiết một kỳ báo cáo
      description: 'report_id.


        Nội dung report đã publish là BẤT BIẾN (I05); delivery không đổi nó (B01). Response mang `report_build_id` (UUIDv4,
        `UNIQUE(owner_id, report_build_id)` — khóa idempotency của `report.publish`, F-A1R2-01) và `abort_reason` (enum
        `empty_period | tag_version_stale | embedding_generation_mismatch | cas_conflict | builder_failure | cancelled`,
        NOT NULL khi và chỉ khi `status=''aborted''`). Hai cột do PC02 FIX3 khai trên ENT-report theo bảng field-level của
        ruling FIX3; hình dạng chi tiết ở `contracts/schemas/report.schema.json` (PC04). `abort_reason=''empty_period''`
        là một kỳ RỖNG HỢP LỆ, KHÔNG phải lỗi — UI không được gộp nó với `builder_failure` (I13).'
      tags:
      - reports
      security:
      - ownerSessionCookie: []
      x-auth-scope: owner_session
      x-owner-module: MOD-report-service
      x-caller-modules:
      - MOD-web-ui
      x-mutation: false
      x-scenario-refs:
      - SC09
      - SC10
      - SC15
      x-requirement-refs:
      - REQ-D19
      - REQ-D53
      - REQ-D54
      - REQ-AC10
      - REQ-P0-07
      x-decision-refs:
      - B14
      - B17
      x-invariant-refs:
      - I05
      - I13
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Ulid'
        description: Report cần xem.
      - $ref: '#/components/parameters/SchemaVersionHeader'
      - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          description: Khối "hướng đang nổi" (nhãn "ứng viên để đọc sâu"), danh sách mục + summary, tag_config_version đã
            dùng, coverage, quality + danh sách pending, ghi chú đợt dừng sớm nếu có.
          content:
            application/json:
              schema:
                $ref: ../schemas/report.schema.json
          headers:
            X-Schema-Version:
              description: Phiên bản schema của response.
              schema:
                $ref: '#/components/schemas/SemVer'
        '401':
          description: 'Lỗi: `UNAUTHORIZED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 'Lỗi: `FORBIDDEN_EDGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 'Lỗi: `NOT_FOUND`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: 'Lỗi: `INTERNAL`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...